+++ Assets/Scripts/idle/upgrades/Handcultivator.cs
@@ -6,7 +6,7 @@ public class HandCultivator : UpgradeBase
 {
     public override float addedMultiplier{get;set;}
     public override float addedBaseClick{get;set;}
-    public override int amountOwned{get;set;}
+    public override float amountOwned{get;set;}
     
     [SerializeField]private float multiplier;
     [SerializeField]private float baseClick;
